Combine user models and modify other models accordingly

+++ backend/controllers/admin.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify, request
 from extensions import db
-from models.admin import Admin
+from models.user import User
 from auth import requires_auth
 
 admins = Blueprint('admins', __name__, url_prefix='/admins')
@@ -9,28 +9,30 @@ admins = Blueprint('admins', __name__, url_prefix='/admins')
 @admins.route('/', methods=['GET'])
 @requires_auth
 def get_admins():
-  admins = db.session.execute(db.select(Admin)).scalars()
+  admins = db.session.scalars(db.select(User).filter_by(is_admin=True))
   return jsonify([admin.dict() for admin in admins])
 
 # Get info for one admin by id
 @admins.route('/<int:id>', methods=['GET'])
 @requires_auth
 def get_admin(id):
-  admin = db.session.get(Admin, id)
-  if admin:
+  admin = db.session.get(User, id)
+  # Admin must exist
+  if admin and admin.is_admin:
     return jsonify(admin.dict())
   return jsonify({'error': 'Admin does not exist'}), 422
 
 # Update info for one admin by id
 @admins.route('/<int:id>', methods=['PUT'])
 @requires_auth
-def update_patient(id):
-  admin = db.session.get(Admin, id)
-  if admin:
+def update_admin(id):
+  admin = db.session.get(User, id)
+  # Admin must exist
+  if admin and admin.is_admin:
     data = request.get_json()
-    admin.first_name = data.get('first_name')
-    admin.last_name = data.get('last_name')
-    admin.email_address = data.get('email_address')
+    # Update admin using keys and values from json
+    for key, value in data.items():
+        setattr(admin, key, value)
     db.session.commit()
     return jsonify(admin.dict())
   return jsonify({'error': 'Admin does not exist'}), 422
@@ -43,10 +45,12 @@ def create_admin():
   first_name = data.get('first_name')
   last_name = data.get('last_name')
   email_address = data.get('email_address')
-  admin = db.session.scalars(db.select(Admin).filter_by(first_name=first_name,last_name=last_name,email_address=email_address)).first()
+  admin = db.session.scalars(db.select(User).filter_by(email_address=email_address)).first()
   if admin:
-    return jsonify({'error': 'Admin already exists'}), 422
-  admin = Admin(first_name=first_name,last_name=last_name,email_address=email_address)
+    if admin.is_admin:
+      return jsonify({'error': 'Admin already exists'}), 422
+    return jsonify({'error': 'User already exists'}), 422
+  admin = User(first_name=first_name,last_name=last_name,email_address=email_address,is_admin=True)
   db.session.add(admin)
   db.session.commit()
   return jsonify(admin.dict())
@@ -55,8 +59,8 @@ def create_admin():
 @admins.route('/<int:id>', methods=['DELETE'])
 @requires_auth
 def delete_admin(id):
-  admin = db.session.get(Admin, id)
-  if admin:
+  admin = db.session.get(User, id)
+  if admin and admin.is_admin:
     db.session.delete(admin)
     db.session.commit()
     return jsonify({'message': 'Admin deleted successfully'})
